+++ Localization/enUS.PluginStrings.lua
@@ -1,0 +1,50 @@
+-- [[ Exported at 2023-01-19 17-25-20 ]] --
+-- [[ This code is automatically generated as an export from ]] --
+-- [[ an SQLite database and is not meant for manual edit. ]] --
+
+-- [[ Namespaces ]] --
+local _, addon = ...;
+addon.PluginStrings = {};
+addon.PluginStrings.enUS = {};
+local pluginStrings = addon.PluginStrings.enUS;
+
+function pluginStrings.Load(L)
+    L["AchievementsTabFixes"] = "Achievements Tab Fixes";
+    L["AchievementsTabFixes Desc"] = "This plugin provides some fixes for the Achievements Tab, toggleable below.";
+    L["RenameExplorationDragonIslesCategory"] = "Rename Exploration -> Dragon Isles Category to Dragonflight";
+    L["RenameExplorationDragonIslesCategory Desc"] = "Renames the Exploration -> Dragon Isles Category to Dragonflight.";
+    L["Better Wardrobe"] = "Better Wardrobe";
+    L["Better Wardrobe Desc"] = "This plugin fixes the link to the wardrobe that is broken by {betterWardrobe}.\n\nThere are no options.";
+    L["Instance Achievement Tracker"] = "Instance Achievement Tracker";
+    L["Instance Achievement Tracker Desc"] = "This plugin allows the addition of an extra Right Click Menu item for {iat}.";
+    L["IAT Tactics"] = "IAT Tactics";
+    L["Add To Right Click Menu"] = "Add To Right Click Menu";
+    L["Add To Right Click Menu Desc"] = "Add an option to the Right Click Menu that will direct you to tactics if {iat} has data for it.";
+    L["Instance Achievement Tracker PluginText"] = "|T:1:8|t- Shortcut to {plugin_iat} if tactics for the achievement are available ({plugin_iat} is {plugin_enabled}).\n";
+    L["Zygor Guides Viewer"] = "Zygor Guides Viewer";
+    L["Zygor Guides Viewer Desc"] = "This plugin adds the magnifying glasses again to the Blizzard Achievements tab.\n\nThere are no options.";
+    L["ElvUI"] = "ElvUI";
+    L["ElvUI Desc"] = "Each of the options below are controlled by ElvUI and are just informational.\n\nTo change these, go to Game Menu -> ElvUI -> Skins and check the desired options. See each option below for what to check.\n";
+    L["Skin Achievements"] = "Skin Achievements";
+    L["Skin Achievements Desc"] = "Applies the ElvUI skin to the Achievements Window.\n-> Blizzard + Achievements";
+    L["Skin Misc Frames"] = "Skin Misc Frames";
+    L["Skin Misc Frames Desc"] = "Applies the ElvUI skin to the Filter Menu, Right Click Menu and Popup Dialog.\n-> Blizzard + Misc Frames";
+    L["Skin Tooltip"] = "Skin Tooltip";
+    L["Skin Tooltip Desc"] = "Applies the ElvUI skin to the Tooltip.\n-> Blizzard + Tooltip";
+    L["Skin Tutorials"] = "Skin Tutorials";
+    L["Skin Tutorials Desc"] = "Applies the ElvUI skin to the Tutorials.\n-> Blizzard + Tutorials";
+    L["Skin Alert Frames"] = "Skin Alert Frames";
+    L["Skin Alert Frames Desc"] = "Applies the ElvUI skin to the Alert Frames.\n-> Blizzard + Alert Frames";
+    L["Skin Calendar"] = "Skin Calendar";
+    L["Skin Calendar Desc"] = "Applies the ElvUI skin to the Calendar.\n-> Blizzard + Calendar Frame";
+    L["Remove Parchment"] = "Remove Parchment";
+    L["Remove Parchment Desc"] = "Remove the parchment background from the Calendar Days.\n-> Parchment Remover";
+    L["Skin Data Manager"] = "Skin Data Manager";
+    L["Skin Data Manager Desc"] = "Applies the ElvUI skin to the Data Manager Window.\n-> Blizzard + Achievements";
+    L["Skin Ace3"] = "Skin Ace3";
+    L["Skin Ace3 Desc"] = "Applies the ElvUI skin to the Options.\n-> Ace3";
+    L["Fix World Map Button"] = "Fix World Map Button";
+    L["Fix World Map Button Desc"] = "When ElvUI Maps -> World Map -> Smaller World Map is enabled, the World Map Button needs fixing";
+    L["Alert System Overwrite Desc"] = "The location settings are not active when ElvUI is enabled. Moving the Loot / Alert Frames anchor will achieve the same result.";
+end
+