+++ Options/Credits.lua
@@ -6,12 +6,12 @@ local credits = options.Credits;
 tinsert(options.OptionsTables, credits);
 
 function credits.RegisterOptionsTable()
-    LibStub("AceConfig-3.0"):RegisterOptionsTable(options.OptionsTable.args.Credits.name, options.OptionsTable.args.Credits);
-    LibStub("AceConfigDialog-3.0"):AddToBlizOptions(options.OptionsTable.args.Credits.name, options.OptionsTable.args.Credits.name, addon.MetaData.Title);
+    LibStub("AceConfig-3.0"):RegisterOptionsTable("Credits", options.OptionsTable.args.Credits);
+    LibStub("AceConfigDialog-3.0"):AddToBlizOptions("Credits", "Credits", addon.MetaData.Title);
 end
 
 function credits.PostLoad()
-    
+
 end
 
 options.OptionsTable.args["Credits"] = {